fix: fix hot reload cache issue
Be careful when importing modules in a function that may be executed
several times -- eg. the useEffect hook. This may cause cache
dismatching in hot reload of webpack dev server.

Also note that the component type changing in hot reload causes screen
blinking, so we render the Markdown component directly in subpage path.